+++ src/Contracts/Toolkit/Identifiers/Identifier.php
@@ -12,9 +12,10 @@ declare(strict_types=1);
 
 namespace CloudCreativity\Modules\Contracts\Toolkit\Identifiers;
 
+use CloudCreativity\Modules\Contracts\Toolkit\Loggable\Contextual;
 use Stringable;
 
-interface Identifier extends Stringable
+interface Identifier extends Stringable, Contextual
 {
     /**
      * Is the identifier the same as the provided identifier?
@@ -37,11 +38,4 @@ interface Identifier extends Stringable
      * @return array-key
      */
     public function key(): string|int;
-
-    /**
-     * Get the value to use when adding the identifier to log context.
-     *
-     * @return mixed
-     */
-    public function context(): mixed;
 }